+++ src/js/main.js
@@ -1,5 +1,6 @@
-import {el, setStyle} from 'redom';
-import {API_URL} from './commons';
+import { el, setStyle } from 'redom';
+import { API_URL } from './commons';
+import { getItem } from "./storage";
 
 export function matchesRegex(str) {
     let reg = /https:\/\/www\.faceit\.com\/([a-z]*)\/csgo\/room\/([a-zA-Z0-9\-]*)$/
@@ -15,24 +16,24 @@ window.onload = () => {
 
 chrome.runtime.onMessage.addListener(
     function (newUrl, sender, callback) {
-        console.log("Detected url change. The new one is: " + newUrl);
+        console.log('Detected url change. The new one is: ' + newUrl);
         validateAndRun();
     });
 
 function validateAndRun() {
-    chrome.storage.sync.get("enableTuscan", function (result) {
-            let enabled = result.enableTuscan;
-
-            if (enabled === "true") {
-                run();
-            } else {
-                console.log("Extension turned off.")
-            }
+    getItem('enableTuscan', result => {
+        let enabled = result.enableTuscan;
+        if (enabled === 'true') {
+            getItem('statsExpanded', statsResult => {
+                run(statsResult.statsExpanded === 'true');
+            });
+        } else {
+            console.log('Faceit lobby extension turned off.')
         }
-    );
+    });
 }
 
-function run() {
+function run(statsExpanded) {
 
     // Find players HTML elements
     let teamMemberElements = document.getElementsByClassName('match-team-member__details__name');
@@ -43,15 +44,15 @@ function run() {
         // Check if extension elements has not been injected to the DOM already
         let tuscanElements = document.getElementsByClassName('tuscan-stats');
         if (tuscanElements.length === 0) {
-            injectApp(teamMemberElements);
+            injectApp(teamMemberElements, statsExpanded);
         }
     } else {
         // Or wait and retry
-        setTimeout(run, 1000);
+        setTimeout(() => run(statsExpanded), 1000);
     }
 }
 
-function injectApp(teamMemberElements) {
+function injectApp(teamMemberElements, statsExpanded) {
 
     // Extract nicknames from player elements
     let nicknames = Array.from(teamMemberElements).map(el => {
@@ -70,11 +71,45 @@ function injectApp(teamMemberElements) {
     let faction2 = nicknames.slice(5);
 
     // Construct table
-    doFaction(faction1, teamMemberElements);
-    doFaction(faction2, teamMemberElements);
+    doFaction(faction1, teamMemberElements, statsExpanded);
+    doFaction(faction2, teamMemberElements, statsExpanded);
 }
 
-function doFaction(faction, teamMemberElements) {
+function setupExpandableDivWithChevron(statsExpanded, table) {
+    const divContainer = el('div');
+    setStyle(divContainer, {
+        borderTop: '1px solid rgb(31, 31, 31)',
+        cursor: 'pointer',
+        height: statsExpanded ? '100%' : '26px'
+    });
+    const chevron = divContainer.appendChild(el('span'));
+    setStyle(chevron, {
+        margin: '6px 8px 10px 0',
+        display: 'inline-block',
+        borderRight: '2px solid #fff',
+        borderBottom: '2px solid #fff',
+        width: '10px',
+        height: '10px',
+        transform: statsExpanded ? 'rotate(45deg)' : 'rotate(-135deg)',
+        float: 'right'
+    });
+    divContainer.appendChild(chevron);
+    divContainer.appendChild(table);
+    divContainer.addEventListener('click', () => {
+        if (table.style.display === 'none') {
+            chevron.style.transform = 'rotate(45deg)';
+            divContainer.style.height = '100%';
+            table.style.display = null;
+        } else {
+            chevron.style.transform = 'rotate(-135deg)';
+            divContainer.style.height = '26px';
+            table.style.display = 'none';
+        }
+    });
+    return divContainer;
+}
+
+function doFaction(faction, teamMemberElements, statsExpanded) {
 
     // Prepare query param
     let query = queryParam(faction);
@@ -94,14 +129,16 @@ function doFaction(faction, teamMemberElements) {
             mapStatsList.forEach(stat => {
 
                 // Create table element:
-                let table = el('table', {className: "tuscan-stats text-sm"});
+                let table = el('table', { className: 'tuscan-stats text-sm' });
 
                 setStyle(table, {
                     width: '100%',
                     color: '#777',
                     backgroundColor: '#181818',
                     marginBottom: '10px',
-                    borderTop: '1px solid #1f1f1f'
+                    borderTop: '1px solid #1f1f1f',
+                    transition: 'all 1s ease',
+                    display: statsExpanded ? null : 'none'
                 });
 
                 // Define columns
@@ -119,10 +156,12 @@ function doFaction(faction, teamMemberElements) {
                     table.appendChild(tr);
                 });
 
+                const divContainer = setupExpandableDivWithChevron(statsExpanded, table);
+
                 // Append constructed table to correlated HTML node
                 Array.from(teamMemberElements).filter(el => el.querySelector('strong').innerText === stat.nickname)
                     .map(el => {
-                        el.parentNode.parentNode.parentNode.firstElementChild.appendChild(table);
+                        el.parentNode.parentNode.parentNode.firstElementChild.appendChild(divContainer);
                         return el
                     });
             });
@@ -135,7 +174,7 @@ function getTuscanLink(nickname) {
         title: `${nickname} TUSCAN stats`,
         href: `https://tuscan.pro/player/${nickname}`,
         target: '_blank'
-    }, "Tuscan Stats");
+    }, 'Tuscan Stats');
 }
 
 function queryParam(nicknames) {